requirements.txt
streamlit==1.55.0
langchain==1.2.12
langchain-google-genai==4.2.1
neo4j==6.1.0
python-dotenv==1.2.2
